making pages less chunky and other small changes
